#138 refactor: dto 디렉토리 구조 변경
